pyproject.toml
[project]
name = "age-bulk-insert-sandbox"
version = "0.1.0"
description = "Benchmark different bulk insertion strategies for Apache AGE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "apache-age-python>=0.0.7",
    "rich>=13.9.0",
    "matplotlib>=3.10.8",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
